Copy the mapping for the current index, to a new index.
Each night, we see the indexing rates for Elasticsearch drop sharply for several
hours. This always happens as new dynamic fields are added and the cluster
master needs to sync these.
With hundreds of nodes, this is a slow process and causes significant lag on
ingestion.

This script will copy the mapping for a type from today's index into tomorrow's.
If you have hundreds or thousands of dynamic fields, it can make a significant
difference to your indexing rates as you start populating a new index.